move function compare_diff_files from the test modules
